minor renaming.
