+++ src/main/java/com/fixspot/backendv1/controllers/ReporterController.java
@@ -7,7 +7,7 @@ import com.fixspot.backendv1.dto.responseDtos.ReporterDetailsResponse;
 import com.fixspot.backendv1.entities.UserEntity;
 import com.fixspot.backendv1.generalUtil.ResultWrapper;
 import com.fixspot.backendv1.generalUtil.Routes;
-import com.fixspot.backendv1.service.services.UserService;
+import com.fixspot.backendv1.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
