+++ src/main/java/com/fixspot/backendv1/controllers/UserAuthController.java
@@ -4,7 +4,7 @@ import com.fixspot.backendv1.auth.CustomUserDetailServiceImpl;
 import com.fixspot.backendv1.auth.JwtService;
 import com.fixspot.backendv1.dto.common.RegisterUserRequest;
 import com.fixspot.backendv1.generalUtil.Routes;
-import com.fixspot.backendv1.service.services.UserService;
+import com.fixspot.backendv1.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
