+++ src/main/java/com/fixspot/backendv1/exception/exceptions/ExpiredTokenException.java
@@ -1,7 +1,6 @@
 package com.fixspot.backendv1.exception.exceptions;
 
 public class ExpiredTokenException extends RuntimeException {
-
     public ExpiredTokenException(String message) {
         super(message);
     }
