+++ src/main/java/com/fixspot/backendv1/exception/exceptions/UserExistsException.java
@@ -4,7 +4,4 @@ public class UserExistsException extends RuntimeException {
     public UserExistsException() {
         super("User Already Exists");
     }
-    public UserExistsException(String message) {
-        super(message);
-    }
 }
